Add auth

+++ src/components/MapList.tsx
@@ -1,6 +1,7 @@
 import React, { useState } from "react";
 import { MapData } from "../types/types";
 import { FaEdit, FaTrash } from "react-icons/fa"; // Import icons for edit and delete
+import { handleGoogleSignIn } from "../util/handleGoogleSignIn";
 
 interface MapListProps {
   maps: MapData[];
@@ -9,6 +10,7 @@ interface MapListProps {
   selectedMapId: string | null;
   setSelectedMapId: (id: string | null) => void;
   deleteMap: (id: string) => void;
+  user: any;
 }
 
 const MapList: React.FC<MapListProps> = ({
@@ -18,41 +20,52 @@ const MapList: React.FC<MapListProps> = ({
   selectedMapId,
   setSelectedMapId,
   deleteMap,
+  user,
 }) => {
   const [newMapName, setNewMapName] = useState("");
   const [editMode, setEditMode] = useState(false); // State for edit mode
 
   const handleAddMap = () => {
-    if (newMapName.trim() === "") return; // Prevent adding maps with empty names
+    if (user) {
+      if (newMapName.trim() === "") return; // Prevent adding maps with empty names
 
-    // Check if the map name is unique
-    const isDuplicate = maps.some((map) => map.name === newMapName);
-    if (isDuplicate) {
-      alert("Map name must be unique.");
-      return;
+      // Check if the map name is unique
+      const isDuplicate = maps.some((map) => map.name === newMapName);
+      if (isDuplicate) {
+        alert("Map name must be unique.");
+        return;
+      }
+
+      addMap(newMapName);
+      setNewMapName("");
+    } else {
+      handleGoogleSignIn();
     }
-
-    addMap(newMapName);
-    setNewMapName("");
   };
 
   const handleDeleteMap = (id: string) => {
-    // Confirm deletion (optional step)
-    const confirmDelete = window.confirm("Are you sure you want to delete this map?");
-    if (confirmDelete) {
-      deleteMap(id);
-      
-      // Select the first map if there are remaining maps
-      if (maps.length > 1) {
-        const remainingMaps = maps.filter(map => map.id !== id);
-        selectMap(remainingMaps[0].id);
-      } else if (maps.length === 1) {
-        // If there is only one map left, select it
-        selectMap(maps[0].id);
-      } else {
-        // No maps left
-        setSelectedMapId(null);
+    if (user) {
+      // Confirm deletion (optional step)
+      const confirmDelete = window.confirm(
+        "Are you sure you want to delete this map?"
+      );
+      if (confirmDelete) {
+        deleteMap(id);
+
+        // Select the first map if there are remaining maps
+        if (maps.length > 1) {
+          const remainingMaps = maps.filter((map) => map.id !== id);
+          selectMap(remainingMaps[0].id);
+        } else if (maps.length === 1) {
+          // If there is only one map left, select it
+          selectMap(maps[0].id);
+        } else {
+          // No maps left
+          setSelectedMapId(null);
+        }
       }
+    } else {
+      handleGoogleSignIn();
     }
   };
 
@@ -97,7 +110,7 @@ const MapList: React.FC<MapListProps> = ({
                     className="text-red-500 ml-2 hover:text-red-700 focus:outline-none"
                     onClick={() => handleDeleteMap(map.id)}
                   >
-                    <FaTrash className="text-red-400 hover:text-red-500"/>
+                    <FaTrash className="text-red-400 hover:text-red-500" />
                   </button>
                 )}
               </li>
